Added terminate signal
Now when a connection is closed using ctrl + c, a string will be sent via serial with "elf2uf2-term\n\r"

This is powerful as now, with this feature, the micro controller can choose to respond to this string by entering boot select. No more plugging and unplugging the microcontroller while holding bootsel.